Single pass is working
Right now there is a behavior that I don't like in the code.
It works perfectly but is fragile and I need to restructure.
The problem is that I need the redux state to automatically move to a
different view under certain resultant states.
I think the best way to accomplish this will be a middleware that looks
at the state after an action and dispatch a view change action
if necessary.  That way it is still another action, but stays automatic.